+++ src/main/java/com/sprint/mission/discodeit/dto/data/MessageDto.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.data;
 
 import java.time.Instant;
 import java.util.List;
