+++ src/main/java/com/sprint/mission/discodeit/dto/request/CreateMessageRequest.java
@@ -1,6 +1,5 @@
-package com.sprint.mission.discodeit.dto.message;
+package com.sprint.mission.discodeit.dto.request;
 
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.util.List;
