+++ src/main/java/com/sprint/mission/discodeit/dto/request/CreatePrivateChannelRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.channel;
+package com.sprint.mission.discodeit.dto.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
