+++ src/main/java/com/sprint/mission/discodeit/dto/request/CreateReadStatusRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.readstatus;
+package com.sprint.mission.discodeit.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.Instant;
