+++ src/main/java/com/sprint/mission/discodeit/dto/request/UpdateChannelRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.channel;
+package com.sprint.mission.discodeit.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
