+++ src/main/java/com/sprint/mission/discodeit/dto/request/UpdateMessageRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.message;
+package com.sprint.mission.discodeit.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
