+++ src/main/java/com/sprint/mission/discodeit/dto/request/UpdateUserRequest.java
@@ -1,13 +1,10 @@
-package com.sprint.mission.discodeit.dto.user;
+package com.sprint.mission.discodeit.dto.request;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 
 @Builder
 public record UpdateUserRequest(
-    @NotBlank
     String newUsername,
-    @NotBlank
     String newEmail,
     String newPassword
 ) {
