+++ src/main/java/com/sprint/mission/discodeit/dto/request/UpdateUserStatusRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.userstatus;
+package com.sprint.mission.discodeit.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.Instant;
