+++ src/main/java/com/sprint/mission/discodeit/mapper/MessageMapper.java
@@ -1,11 +1,9 @@
 package com.sprint.mission.discodeit.mapper;
 
-import com.sprint.mission.discodeit.dto.MessageDto;
+import com.sprint.mission.discodeit.dto.data.MessageDto;
 import com.sprint.mission.discodeit.entity.Message;
-import lombok.RequiredArgsConstructor;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.springframework.stereotype.Component;
 
 @Mapper(componentModel = "spring", uses = {UserMapper.class, BinaryContentMapper.class})
 public interface MessageMapper {
