+++ src/main/java/com/sprint/mission/discodeit/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.mapper;
 
-import com.sprint.mission.discodeit.dto.UserDto;
+import com.sprint.mission.discodeit.dto.data.UserDto;
 import com.sprint.mission.discodeit.entity.User;
 import java.util.List;
 import org.mapstruct.Mapper;
