+++ src/test/java/com/sprint/mission/discodeit/repository/MessageRepositoryTest.java
@@ -17,6 +17,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
+import org.springframework.data.domain.Sort;
 import org.springframework.test.context.ActiveProfiles;
 
 @DataJpaTest
@@ -70,7 +71,7 @@ class MessageRepositoryTest {
     // given
     UUID channelId = channelRepository.findAll().get(0).getId();
     Instant cursor = messageRepository.findAllByChannel_Id(channelId).get(0).getCreatedAt();
-    Pageable pageable = PageRequest.of(0, 10);
+    Pageable pageable = PageRequest.of(0, 10, Sort.by("createdAt").descending());
 
     // when
     Slice<Message> slice = messageRepository.findAllByChannel_IdAndCreatedAtLessThanEqual(channelId, cursor, pageable);
